Project 05_02 Almost finished
When it comes to project #2, I have passed through all questions and applied codes of some kind, but in a similar fashion to the last project, I have made slight oversights in the CSS that I must find in order to resolve the layout to its proper order. That and I need to also add in comments as well. Both of which I plan to do over the weekend.